Record game state at the start of every tick
Save a copy of the last N game states. This should (in theory) allow
for developers to rewind the game, and thus make debugging easier.

Even better would be to also record that ticks input, so that the
game can play itself, and thus make bugs easier to investigate.

Bonus: Tidy up Makefile

 - Split CFLAGS into multiple lines
 - Headline comments